refactor: seperate of concern
