+++ src/views/productManagerView.js
@@ -1,4 +1,4 @@
-import { entryObject } from "../utils/utils.js";
+import { View } from "../core/View";
 
 export function viewInitiator(rootElement, HTML) {
   rootElement.innerHTML = HTML;
@@ -14,36 +14,6 @@ export function viewInitiator(rootElement, HTML) {
     }),
     productInventoryContainerView: new View($productInventoryContainer),
   };
-}
-
-class View {
-  rootElement = null;
-  #dynamicElementSelectors = {};
-
-  constructor(rootElement, dynamicElementSelectors = {}) {
-    this.rootElement = rootElement;
-    this.#dynamicElementSelectors = dynamicElementSelectors;
-    this.#scanDynamicElement();
-  }
-
-  #scanDynamicElement() {
-    entryObject(this.#dynamicElementSelectors).forEach(([name, selector]) => {
-      const element = this.rootElement.querySelector(selector);
-      if (!element) throw new Error(`selector "${selector}" does not exist`);
-      this[name] = element;
-    });
-  }
-
-  attachEvent(eventObject) {
-    entryObject(eventObject).forEach(([name, [eventType, eventCallback]]) => {
-      this[name].addEventListener(eventType, eventCallback);
-    });
-  }
-
-  render(newElement) {
-    this.rootElement.replaceWith(newElement);
-    this.#scanDynamicElement();
-  }
 }
 
 export function getProductManagerHTML() {
